cloud_function/requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
pandas==2.1.3
google-cloud-storage==2.13.0
google-cloud-bigquery==3.13.0
requests==2.31.0
functions-framework==3.4.0
flask>=2.0.0
pyarrow>=14.0.1  # Ajout de pyarrow